+++ src/PieceQuadrilatere.java
@@ -1,8 +1,8 @@
 public class PieceQuadrilatere extends Piece {
-    private int _longueur;
-    private int _largeur;
+    private double _longueur;
+    private double _largeur;
 
-    public PieceQuadrilatere(TypePiece typePiece, String niveau, int uneLongueur, int uneLargeur) {
+    public PieceQuadrilatere(TypePiece typePiece, String niveau, double uneLongueur, double uneLargeur) {
         super(typePiece, niveau);
         this._longueur = uneLongueur;
         this._largeur = uneLargeur;
@@ -16,11 +16,11 @@ public class PieceQuadrilatere extends Piece {
         return result;
     }
 
-    public int get_longueur() {
+    public double get_longueur() {
         return _longueur;
     }
 
-    public int get_largeur() {
+    public double get_largeur() {
         return _largeur;
     }
